Updated Calculator

+++ Labs/calculator/Calculate.h
@@ -29,7 +29,7 @@
             double sub (double a, double b);
             double multiply (double a, double b);
             double divide (double a, double b);
-            int modulo (double a, double b);
+            int modulo (int a, int b);
